Upgrades react and react native and fixes issues with SampleApp due to upgrade
