+++ SampleApp/package.json
@@ -6,8 +6,8 @@
     "start": "react-native run-ios"
   },
   "dependencies": {
-    "react": "^15.1.0",
-    "react-native": "^0.27.2",
+    "react": "^15.3.0",
+    "react-native": "^0.36.1",
     "react-native-modal-picker": "file:../"
   }
 }
